+++ pulse-binding/src/stream.rs
@@ -307,6 +307,7 @@ type EventCb = callbacks::MultiUseCallback<dyn FnMut(String, Proplist),
 /// The state of a stream.
 #[repr(C)]
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(FromPrimitive, ToPrimitive)]
 pub enum State {
     /* NOTE: This enum’s variants and variant values **must** remain identical to the `sys` crate
        (C API) equivalent */