+++ pulse-sys/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [unreleased]
+
+ * Added derive of `FromPrimitive` and `ToPrimitive` from the `num-derive` crate on enums
+
 # 1.14.1 (November 25th, 2020)
 
  * Fixed missing rename of `pa_error_code_t::Io` to `pa_error_code_t::IO` to match change in binding